8905 - Import puvodcu CAM odkazovanych z PevaOriginator

+++ src/transfagent/src/main/java/cz/aron/transfagent/repository/CoreQueueRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import cz.aron.transfagent.domain.CoreQueue;
 
-public interface CoreQueueRepository extends JpaRepository<CoreQueue, Integer> {
+public interface CoreQueueRepository extends JpaRepository<CoreQueue, Long> {
 
 	@EntityGraph(attributePaths = {"apuSource"})
     CoreQueue findFirstByOrderById();
